+++ packages/schemas/jest.config.ts
@@ -2,6 +2,7 @@
 export default {
   displayName: 'schemas',
   preset: '../../jest.preset.js',
+  verbose: true,
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/tsconfig.spec.json',
